requirements.txt
gradio
tensorflow
opencv-python-headless
numpy
Pillow
